ANDROID: verity: really fix android-verity Kconfig
The change "ANDROID: verity: fix android-verity Kconfig dependencies"
relaxed the dependency on DM_VERITY=y to just DM_VERITY, but this is not
correct because there are parts of the verity and dm-mod API that
android-verity is using but which are not exported to modules.

Work around this problem by disallowing android-verity to be built-in
when the dm/verity core is built modularly.

Bug: 72722987
Change-Id: I3cfaa2acca8e4a4b5c459afdddd958ac9f8c1eb3
Signed-off-by: Alistair Strachan <astrachan@google.com>
(cherry picked from commit a9218b60578bfc9870d2d3c08088cc049fbfed1f)
Signed-off-by: Guenter Roeck <groeck@chromium.org>