RISC-V: Align .Lsecondary_park
The ISA mandantes that stvec is 4-byte aligned, but we weren't actually
respecting this for .Lsecondary_park.  This trap vector is never
expected to actually be used unless there's a bug in early in the boot
process of the kernel, which explains why the bug wasn't noticed until
recently.

Thanks to Zong for finding the bug!

CC: zong@andestech.com
Signed-off-by: Palmer Dabbelt <palmer@sifive.com>