dm table: use logical_block_size of underlying device when verifying start and len
It was found that creating a crypt device with 4K sector size ontop of a
device that has a 512b logical_block_size can incorrectly fail:
  # modprobe scsi_debug physblk_exp=3 lowest_aligned=1
  # cat /sys/block/sda/queue/logical_block_size
  512
  # dmsetup create crypt0 --table "0 8192 crypt cipher_null-ecb - 0 /dev/sda 1 1 sector_size:4096"
with the error:
  device-mapper: table: 253:0: start=1 not aligned to h/w logical block size 4096 of sda

The device_area_is_invalid() logic incorrectly checks the starting
sector offset against the logical block size of the device that is being
created.  It needs to be checking the 'start' offset against the
underlying device's logical block size.  The same goes for 'len'
checking.

Cc: stable@vger.kernel.org
Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>