FROMGIT: pinctrl: msm: Configure interrupts as input and gpio mode
When requesting a gpio as an interrupt, we should make sure to mux the
pin as the GPIO function and configure it to be an input so that various
functions or output signals don't affect the interrupt state of the pin.
So far, we've relied on pinmux configurations in DT to handle this, but
let's explicitly configure this in the code so that DT implementers
don't have to get this part right.

Cc: Doug Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
(cherry picked from commit fe2731211e784bff5a738631e5fd48041524239f
 git://git.kernel.org/pub/scm/linux/kernel/git/linusw/linux-pinctrl.git for-next)

BUG=None
TEST=Remove input/mode settings from DT and see it still gets interrupts

Change-Id: I96370ce569df26cd62bed912429c8b63a539ecc2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1105297
Reviewed-by: Stephen Boyd <swboyd@chromium.org>