CHROMIUM: scripts: Account for clang in the kernel config scripts
It appears that upstream now accounts for your compiler version in
your config.  This means that the kernel config that we're using isn't
what's actually in our split config.

Given that the compiler version / options affect the config, we should
enforce working with the split config to use the same compiler version
that the kernel will be compiled on.  This makes it very clear what
our options will be and also allows us to enable config options that
might only be available with certain compilers.

One big downside here is that you'll only be able to normalize the
config if you've got _all_ the cross compilers installed.  You'll also
only be able to normalize inside the chroot.

NOTE: there's some duplication here with the cros-kernel2.eclass and
various overlays in terms of deciding what compiler will be used.  I'm
not sure how to fix this without turning the config normalization into
an ebuild command.

ALSO NOTE: If there is some case where someone will need to compile
the kernel with a different compiler (maybe by specifying USE flags)
things should still be OK.  The kernel config system should adapt the
config at build time.

BUG=chromium:919701
TEST=Normalize the config and see that it looks sane.

Change-Id: I47fc378e4a7a34352a598a3df3d722d189485a5c
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1330711
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>