msm: vidc: Fix broken debugfs creation error checks and error paths
The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

This fixes the kernel panic on boot when debugfs is disabled.

Change-Id: Ia9bc06def83d05fb1c0d54c8f6ff8fad730193de
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>