autofs: add strictexpire mount option
092a53452b (("autofs: take more care to not update last_used on path
walk") helped to (partially) resolve a problem where automounts were not
expiring due to aggressive accesses from user space.

This patch was later reverted because, for very large environments, it
meant more mount requests from clients and when there are a lot of clients
this caused a fairly significant increase in server load.

But there is a need for both types of expire check, depending on use case,
so add a mount option to allow for strict update of last use of autofs
dentrys (which just means not updating the last use on path walk access).

Link: http://lkml.kernel.org/r/154296973880.9889.14085372741514507967.stgit@pluto-themaw-net
Signed-off-by: Ian Kent <raven@themaw.net>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>