Btrfs: check if destination root is read-only for deduplication
Checking if the destination root is read-only was being performed only for
clone operations. Make deduplication check it as well, as it does not make
sense to not do it, even if it is an operation that does not change the
file contents (such as defrag for example, which checks first if the root
is read-only).

As dedupe can cause problems with send that assumes that a read-only
subvolume does not change, the read-only check is added to prevent
breakage (incorrect data sent/deduped, unexpected crashes).

This is a problem introduced by dedupe ioctl when it landed, since send
existed for a longer time (when nothing else was allowed to change
read-only roots, including defrag).

The main usecase being a primary subvolume with RO snapshots over time,
with optional incremental send. That we know is common and widely used.

A problematic usecase that would utilize deduplication over RO snapshots
could be something like a set of subvolumes that have similar data, get
snapshotted or set RO, followed by a dedupe pass.

To support the latter we should go only via the RO->RW, dedupe, RW->RO
way, as this records that there was a change and does not collide with
send assumptions. That this leads to loss of incremental send must be
communicated to the user with possible override.

Signed-off-by: Filipe Manana <fdmanana@suse.com>
Reviewed-by: David Sterba <dsterba@suse.com>
[ copied rationale from the mail discussions ]
Signed-off-by: David Sterba <dsterba@suse.com>