smb3: make default i/o size for smb3 mounts match what server's prefer
We negotiate rsize and wsize on mounts (and can be overridden by user) -
typically is 4MB.  This is much better (and not just for long latency
network connections) in most use cases for SMB3 than the default I/O
size (which ends up being 128K for cp and smaller than scp).

By changing inode->blocksize from 32K (which was perhaps
reasonable for SMB1/CIFS) to the value negotiated with the server
which is usually 4MB, this significantly increases performance for
the noncached case, and slightly increases it for the cached case.

Signed-off-by: Steve French <stfrench@microsoft.com>
Reviewed-by: Ronnie Sahlberg <lsahlber@redhat.com>
CC: Stable <stable@vger.kernel.org>