mm/hmm: potential deadlock in nonblocking code
There is a deadlock bug when these functions are used in nonblocking mode.

The else side of the if/else statement is only meant to be taken in when
the code is used in blocking mode.  But, unfortunately, the way the code
is now, if we're in non-blocking mode and we succeed in taking the lock
then we do the else statement.  The else side tries to take lock a second
time which results in a deadlock.

Link: http://lkml.kernel.org/r/20190204182304.GA8756@kadam
Fixes: a3402cb621c1 ("mm/hmm: improve driver API to work and wait over a range")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Reviewed-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Matthew Wilcox <willy@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>