mmc: core: Fix a bug when max_discard is 0
The original purpose of the code, is to choose max_trim as the final
parameter, rather than max_discard if max_trim is less than max_discard.

However, in an uncommon case max_discard may be 0 because of overflowing
the maximum threshold. In this case, we should choose max_trim instead of
max_discard, so fix the code to do that.

Signed-off-by: Jiong Wu <Lohengrin1024@gmail.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>