CHROMIUM: arm64: dts: mt8183-kukui: Add more backlight levels
Current backlight settings are a simple linear mapping from 0-16
to 0-1023, which leads to not-so-smooth backlight level transitions
(despite the fact the that underlying PWM can support 1024 levels).

According to the datasheet of RT8532, we should have pwm frequency
less than or equal to 2kHz to support 1024 brightness levels. As for
mt6370, it may also depend on sampling rate, but 2kHz seems fine for
now.

Let's extend the mapping from 0-1023 to 0-1023 and decrease the pwm
frequency to 2kHz.

BUG=b:123845266
TEST=Boot kukui, cat /sys/class/backlight/backlight_lcd0/max_brightness
     => 1023, backlight transition looks smooth.

Change-Id: Ic57982b12e9166ec9ec2f975655103642af023ba
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Signed-off-by: Claire Chang <tientzu@chromium.org>
Signed-off-by: Eddie Huang <eddie.huang@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/1451676
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>