CIFS: Reset current mid to next value in sequence on send failures
When we hit failures during constructing MID or sending the PDU
through the network, we end up not using the message ID assigned
to the packet. The next SMB packet will skip that message ID and
continue with the next one. This behavior may lead to a server not
granting us credits until we use the skipped ID. Fix this by
reverting the ID to the original value if any errors occur before
we push the packet through the network stack.

This patch fixes the generic/310 test from the xfs-tests.

Signed-off-by: Pavel Shilovsky <pshilov@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>