huegtlbfs-fix-races-and-page-leaks-during-migration-update
>>  	spin_unlock(ptl);
>> +
>> +	/* May already be set if not newly allocated page */
>> +	set_page_huge_active(page);
>> +

This is wrong.  We need to only set_page_huge_active() for newly allocated
pages.  Why?  We could have got the page from the pagecache, and it could
be that the page is !page_huge_active() because it has been isolated for
migration.  Therefore, we do not want to set it active here.

I have also found another race with migration when removing a page from
a file.  When a huge page is removed from the pagecache, the page_mapping()
field is cleared yet page_private continues to point to the subpool until
the page is actually freed by free_huge_page().  free_huge_page is what
adjusts the counts for the subpool.  A page could be migrated while in this
state.  However, since page_mapping() is not set the hugetlbfs specific
routine to transfer page_private is not called and we leak the page count
in the filesystem.  To fix, check for this condition before migrating a huge
page.  If the condition is detected, return EBUSY for the page.

Both issues are addressed in the updated patch below.

Sorry for the churn.  As I find and fix one issue I seem to discover another.
There is still at least one more issue with private pages when COW comes into
play.  I continue to work that.  I wanted to send this patch earlier as it
is pretty easy to hit the bugs if you try.  If you would prefer another
approach, let me know.

Link: http://lkml.kernel.org/r/7534d322-d782-8ac6-1c8d-a8dc380eb3ab@oracle.com
Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: "Kirill A . Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>