btrfs: merge btrfs_setxattr and do_setxattr
Now btrfs_setxattr() is a very small function which checks for readonly
FS and redirects the call to do_setxattr(). So instead move that checks
to the parent functions and call do_setxattr() directly.  Merge original
btrfs_setxattr and do_setxattr() to btrfs_setxattr().  Note the
arguments of original do_setxattr() and original btrfs_setxattr() are
same, so the diff obliterates the changes as described above.

Signed-off-by: Anand Jain <anand.jain@oracle.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>