FROMLIST: media: mtk-vcodec: vdec: move stateful ops into their own file
We are planning to add support for stateless formats to this driver.
Part of the driver will be shared between stateful and stateless
formats, but a few ops need to be specialized for both. Extract the
stateful part of the driver and move it into its own file, accessible
through ops that the common driver parts can call.

This patch only moves code around and introduces a set of abstractions ;
behavior or the driver should not be changed in any way.

Signed-off-by: Yunfei Dong <yunfei.dong@mediatek.com>
[acourbot: refactor, cleanup and split]
Co-developed-by: Alexandre Courbot <acourbot@chromium.org>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

(am from https://patchwork.kernel.org/patch/11135069/)

BUG=b:123551776
TEST=Kernel builds for Kukui.

Change-Id: I12375c79af8b5a6a8c2e609ec0894fea32eb754a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1798312
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>