XArray: Add xas_replace
Jan found a performance regression between the radix tree and the XArray
when truncating a file which is almost entirely swap entries.  Most of
that regression was due to the tags being reset when removing the swap
entry from the tree.  This is unnecessary because the tags are always
clear for swap entries.  Use xas_replace() in parts of the pagecache
where we know it's safe to do so.

Signed-off-by: Matthew Wilcox <willy@infradead.org>