btrfs: Defer setting new inode mode until after do_set_acl succeeds
Currently a reference to inode->i_mode is passed directly to
posix_acl_update_mode when setting an ACL which results in the inode's
mode always being changed. In case of errors (e.g. in do_set_acl or
even starting a transaction) the old mode needs to be re-assigned to
->i_mode. This mode recovery needs to be done if do_set_acl or
btrfs_start_transaction fail.

Simplify it by setting the new mode to a temporary variable which is
assigned to inode->i_mode's only when do_set_acl succeeds. This covers
both failure cases explained above.

Signed-off-by: Nikolay Borisov <nborisov@suse.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>