lib: bitmap_parselist: don't calculate length of the input string
Patch series "lib: rework bitmap_parselist and tests", v2.

bitmap_parselist has been evolved from a pretty simple idea for long and
now lacks for refactoring.  It is not structured, has nested loops and a
set of opaque-named variables.

Things are more complicated than they may be because bitmap_parselist() is
a part of user interface, and its behavior should not change.

In this patchset
 - __bitmap_parselist() is reworked (patches 2 and 3);
 - time measurement in test_bitmap_parselist switched to ktime_get
   (patch 4);
 - new tests introduced (patch 5), and
 - bitmap_parselist_user() testing enabled with the same testset as
   bitmap_parselist() (patch 6).

Patches 1 and 4 are fixes and may be applied separately.

v1: https://lkml.org/lkml/2018/12/23/50
v2: https://www.spinics.net/lists/kernel/msg3048976.html

This patch (of 6):

bitmap_parselist() calculates length of the input string before passing it
to the __bitmap_parselist().  But the end-of-line condition is checked for
every character in __bitmap_parselist() anyway.  So doing it in wrapper is
a simple waste of time.

Link: http://lkml.kernel.org/r/20190325210748.6571-2-ynorov@marvell.com
Signed-off-by: Yury Norov <ynorov@marvell.com>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Kees Cook <keescook@chromium.org>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>